+++ app/assets/javascripts/Students/index.js
@@ -1,5 +1,13 @@
 var modalNotesGroup = null;
 
-$(document).ready(function () {
-  modalNotesGroup = new ModalMarkus("#notes_dialog");
-});
+(function () {
+  const domContentLoadedCB = function () {
+    modalNotesGroup = new ModalMarkus("#notes_dialog");
+  };
+
+  if (document.readyState === "loading") {
+    document.addEventListener("DOMContentLoaded", domContentLoadedCB);
+  } else {
+    domContentLoadedCB();
+  }
+})();