+++ app/assets/javascripts/check_timeout.js
@@ -1,5 +1,9 @@
-$(document).ready(() => {
-  setInterval(() => {
-    $.get(Routes.check_timeout_main_index_path());
-  }, 120000);
-});
+(function () {
+  const domContentLoadedCB = () => {
+    setInterval(() => {
+      $.get(Routes.check_timeout_main_index_path());
+    }, 120000);
+  };
+
+  document.addEventListener("DOMContentLoaded", domContentLoadedCB);
+})();